Update tools.py
making all future database queries as a class, with functions to get results/error/length

for in the future, if we need to convert into a different datastore (i.e. mongoDB, PSQL, cassandra), we just have to changes the queries and class, not the whole script.

changes how ```column``` is handled, preventing duplicating the same code over and over again. (can lead to errors if not both are the same)